added skeleton for troubleshooting
